Deleted StateService and used PostService instead
