+++ src/app/components/feed/feed.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { StateService } from 'src/app/services/state.service';
+import { PostService } from 'src/app/services/post.service';
 import { Post } from 'src/types/posts';
 
 @Component({
@@ -11,11 +11,9 @@ export class FeedComponent{
 
   public _posts: Post[];
 
-  constructor(private _stateService: StateService) { 
-    this._stateService.posts$
+  constructor(private _postService: PostService) { 
+    this._postService.posts$
       .subscribe(posts => this._posts = posts);
   }
-
-
 
 }
