--- src/app/services/state.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { StateService } from './state.service';
-
-describe('StateService', () => {
-  let service: StateService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(StateService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
